fix a few warnings (format strings, unused variable)
During compilation gcc complains about the following:

perl.c:4970: warning: format '%u' expects argument of type 'unsigned int', but argument 2 has type 'U32' [-Wformat=]
perl.c:5075: warning: format '%u' expects argument of type 'unsigned int', but argument 2 has type 'I32' [-Wformat=]
mg.c:1972: warning: format '%ld' expects argument of type 'long int', but argument 2 has type 'ssize_t' [-Wformat=]
pp_ctl.c:2610: warning: unused variable 'mark' [-Wunused-variable]
regexec.c:2275: warning: format '%ld' expects argument of type 'long int', but argument 3 has type 'int' [-Wformat=]

This patch fixes all of them.

Tony: warning: unused variable 'mark' was fixed in 481c819b

+++ perl.c
@@ -4902,7 +4902,7 @@ Perl_my_exit(pTHX_ U32 status)
     }
     if (PL_exit_flags & PERL_EXIT_WARN) {
 	PL_exit_flags |= PERL_EXIT_ABORT; /* Protect against reentrant calls */
-	Perl_warn(aTHX_ "Unexpected exit %u", status);
+	Perl_warn(aTHX_ "Unexpected exit %lu", (unsigned long)status);
 	PL_exit_flags &= ~PERL_EXIT_ABORT;
     }
     switch (status) {
@@ -5007,7 +5007,7 @@ Perl_my_failure_exit(pTHX)
     }
     if (PL_exit_flags & PERL_EXIT_WARN) {
 	PL_exit_flags |= PERL_EXIT_ABORT; /* Protect against reentrant calls */
-	Perl_warn(aTHX_ "Unexpected exit failure %u", PL_statusvalue);
+	Perl_warn(aTHX_ "Unexpected exit failure %ld", (long)PL_statusvalue);
 	PL_exit_flags &= ~PERL_EXIT_ABORT;
     }
     my_exit_jump();
